tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/execution-timing/040.html

<!DOCTYPE html>
<html><head>
        <title> scheduler: IFRAMEs added with DOM (innerHTML), javascript: URL</title>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
        <script src="testlib/testlib.js"></script>
</head>
<body>

        <div id="log">FAILED (This TC requires JavaScript enabled)</div>
        <div></div>
        <script>log('inline script #1');
                document.getElementsByTagName('div')[1].innerHTML = '<iframe src="javascript:parent.log(\'JS URL\');\'<html><script>parent.log(\\\'frame script\\\')<\/script></html>\'"></iframe>';
                log('end script #1');
        </script>

        <script type="text/javascript">
        log( 'inline script #2' );
        var t = async_test()

        function test() {
                assert_any(assert_array_equals, eventOrder, [
                  ['inline script #1', 'end script #1', 'JS URL', 'inline script #2', 'frame script'],
                  /* the following combination seems quite unlikely? */
                  ['inline script #1', 'end script #1', 'JS URL', 'frame script', 'inline script #2'],
                  ['inline script #1', 'end script #1', 'inline script #2', 'JS URL', 'frame script']]);
                t.done();
        }
        onload = t.step_func(function(){
                setTimeout(t.step_func(test), 200);
        });
        </script>

</body></html>
