third_party/WebKit/LayoutTests/external/wpt/html/semantics/scripting-1/the-script-element/execution-timing/040-expected.txt

This is a testharness.js-based test.
FAIL  scheduler: IFRAMEs added with DOM (innerHTML), javascript: URL assert_array_equals: property 1, expected "end script #1" but got "JS URL"

assert_array_equals: property 1, expected "end script #1" but got "JS URL"

assert_array_equals: property 1, expected "end script #1" but got "JS URL"
Harness: the test ran to completion.
